+++ backend/package.json
@@ -19,6 +19,7 @@
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "fs": "^0.0.1-security",
+    "jsonwebtoken": "^9.0.2",
     "localStorage": "^1.0.4",
     "mongoose": "^8.2.4",
     "multer": "^1.4.5-lts.1",
